# backend/app/api/v1/subscription.py
"""
Subscription API endpoints

This module provides API endpoints for subscription management,
payment processing, and quota operations.
"""

from fastapi import APIRouter, HTTPException, Depends, status
from fastapi.security import HTTPBearer
from pydantic import BaseModel
from typing import Optional, List, Dict, Any
from decimal import Decimal
import logging

from app.core.logging import get_logger

logger = get_logger(__name__)
router = APIRouter()
security = HTTPBearer()


# Pydantic models
class SubscriptionPlan(BaseModel):
    """Subscription plan model."""
    plan_type: str
    name: str
    monthly_quota: int
    price: Decimal
    duration_days: Optional[int]
    features: List[str]
    is_popular: bool = False


class CreateOrderRequest(BaseModel):
    """Create subscription order request."""
    plan_type: str
    payment_method: str = "wechat"


class PaymentRequest(BaseModel):
    """Payment request model."""
    order_no: str
    payment_method: str
    return_url: Optional[str] = None


class SubscriptionResponse(BaseModel):
    """Subscription response model."""
    plan_type: str
    plan_name: str
    status: str
    monthly_quota: int
    used_quota: int
    remaining_quota: int
    start_date: str
    end_date: Optional[str]
    auto_renew: bool
    is_active: bool


class OrderResponse(BaseModel):
    """Order response model."""
    order_no: str
    plan_type: str
    amount: Decimal
    currency: str
    payment_status: str
    payment_method: str
    created_at: str
    expires_at: Optional[str]


@router.get("/plans", response_model=List[SubscriptionPlan])
async def get_subscription_plans():
    """
    Get available subscription plans.
    
    This endpoint returns all available subscription plans
    with pricing and feature information.
    """
    try:
        # TODO: Implement plans retrieval logic
        # 1. Get plans from configuration or database
        # 2. Include current promotions if any
        # 3. Format response
        
        # Mock response
        mock_plans = [
            SubscriptionPlan(
                plan_type="free",
                name="免费版",
                monthly_quota=1,
                price=Decimal("0.00"),
                duration_days=None,
                features=["每月1次翻译", "基础语音识别"],
                is_popular=False
            ),
            SubscriptionPlan(
                plan_type="basic_vip",
                name="基础VIP",
                monthly_quota=10,
                price=Decimal("9.90"),
                duration_days=30,
                features=["每月10次翻译", "高质量语音识别", "优先处理"],
                is_popular=True
            ),
            SubscriptionPlan(
                plan_type="premium_vip",
                name="高级VIP",
                monthly_quota=100,
                price=Decimal("29.90"),
                duration_days=30,
                features=["每月100次翻译", "最高质量语音识别", "最快处理", "客服支持"],
                is_popular=False
            )
        ]
        
        return mock_plans
        
    except Exception as e:
        logger.error(f"Get subscription plans error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取订阅方案失败"
        )


@router.get("/current", response_model=SubscriptionResponse)
async def get_current_subscription(current_user: dict = Depends(security)):
    """
    Get current user's subscription information.
    
    This endpoint returns the user's active subscription
    details including quota and status.
    """
    try:
        # TODO: Implement current subscription retrieval
        # 1. Get user's active subscription
        # 2. Calculate remaining quota
        # 3. Include next billing date
        # 4. Format response
        
        # Mock response
        return SubscriptionResponse(
            plan_type="free",
            plan_name="免费版",
            status="active",
            monthly_quota=1,
            used_quota=0,
            remaining_quota=1,
            start_date="2024-01-01T00:00:00Z",
            end_date=None,
            auto_renew=False,
            is_active=True
        )
        
    except Exception as e:
        logger.error(f"Get current subscription error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取当前订阅失败"
        )


@router.post("/orders", response_model=OrderResponse)
async def create_subscription_order(
    request: CreateOrderRequest,
    current_user: dict = Depends(security)
):
    """
    Create subscription upgrade order.
    
    This endpoint creates a payment order for subscription
    upgrade or renewal.
    """
    try:
        # TODO: Implement order creation logic
        # 1. Validate plan type
        # 2. Calculate pricing
        # 3. Create order record
        # 4. Set expiration time
        # 5. Return order details
        
        # Mock response
        mock_order_no = f"SUB{int(__import__('time').time())}1234"
        
        return OrderResponse(
            order_no=mock_order_no,
            plan_type=request.plan_type,
            amount=Decimal("9.90") if request.plan_type == "basic_vip" else Decimal("29.90"),
            currency="CNY",
            payment_status="pending",
            payment_method=request.payment_method,
            created_at="2024-01-15T10:00:00Z",
            expires_at="2024-01-15T12:00:00Z"
        )
        
    except Exception as e:
        logger.error(f"Create subscription order error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="创建订阅订单失败"
        )


@router.post("/pay/wechat", response_model=dict)
async def create_wechat_payment(
    request: PaymentRequest,
    current_user: dict = Depends(security)
):
    """
    Create WeChat Pay payment parameters.
    
    This endpoint generates WeChat Pay parameters for
    completing subscription payment.
    """
    try:
        # TODO: Implement WeChat Pay integration
        # 1. Validate order exists and belongs to user
        # 2. Generate WeChat Pay parameters
        # 3. Create payment record
        # 4. Return payment parameters
        
        return {
            "success": True,
            "data": {
                "payment_params": {
                    "appId": "mock_wechat_appid",
                    "timeStamp": "1642234567",
                    "nonceStr": "mock_nonce_string",
                    "package": "prepay_id=mock_prepay_id",
                    "signType": "MD5",
                    "paySign": "mock_signature"
                },
                "order_no": request.order_no
            },
            "message": "支付参数生成成功"
        }
        
    except Exception as e:
        logger.error(f"Create WeChat payment error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="创建微信支付失败"
        )


@router.post("/pay/alipay", response_model=dict)
async def create_alipay_payment(
    request: PaymentRequest,
    current_user: dict = Depends(security)
):
    """
    Create Alipay payment parameters.
    
    This endpoint generates Alipay parameters for
    completing subscription payment.
    """
    try:
        # TODO: Implement Alipay integration
        # 1. Validate order exists and belongs to user
        # 2. Generate Alipay parameters
        # 3. Create payment record
        # 4. Return payment parameters
        
        return {
            "success": True,
            "data": {
                "payment_params": {
                    "app_id": "mock_alipay_appid",
                    "method": "alipay.trade.app.pay",
                    "charset": "utf-8",
                    "sign_type": "RSA2",
                    "timestamp": "2024-01-15 10:00:00",
                    "version": "1.0",
                    "notify_url": "https://api.example.com/payment/alipay/notify",
                    "sign": "mock_alipay_signature"
                },
                "order_no": request.order_no
            },
            "message": "支付参数生成成功"
        }
        
    except Exception as e:
        logger.error(f"Create Alipay payment error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="创建支付宝支付失败"
        )


@router.get("/orders", response_model=List[OrderResponse])
async def get_payment_history(
    page: int = 1,
    limit: int = 20,
    current_user: dict = Depends(security)
):
    """
    Get user's payment order history.
    
    This endpoint returns a paginated list of user's
    payment orders and their status.
    """
    try:
        # TODO: Implement payment history retrieval
        # 1. Query user's payment orders
        # 2. Apply pagination
        # 3. Format response
        
        # Mock response
        mock_orders = [
            OrderResponse(
                order_no="SUB16422345671234",
                plan_type="basic_vip",
                amount=Decimal("9.90"),
                currency="CNY",
                payment_status="paid",
                payment_method="wechat",
                created_at="2024-01-15T10:00:00Z",
                expires_at=None
            )
        ]
        
        return mock_orders
        
    except Exception as e:
        logger.error(f"Get payment history error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取支付历史失败"
        )


@router.post("/cancel", response_model=dict)
async def cancel_subscription(current_user: dict = Depends(security)):
    """
    Cancel current subscription.
    
    This endpoint cancels the user's active subscription.
    Access remains until the end of billing period.
    """
    try:
        # TODO: Implement subscription cancellation
        # 1. Get user's active subscription
        # 2. Update status to cancelled
        # 3. Disable auto-renewal
        # 4. Send confirmation
        
        return {
            "success": True,
            "message": "订阅已取消，您可以在当前计费周期结束前继续使用所有功能"
        }
        
    except Exception as e:
        logger.error(f"Cancel subscription error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="取消订阅失败"
        )


@router.post("/auto-renew", response_model=dict)
async def toggle_auto_renew(
    enabled: bool,
    current_user: dict = Depends(security)
):
    """
    Toggle subscription auto-renewal.
    
    This endpoint enables or disables automatic subscription
    renewal for the user's active subscription.
    """
    try:
        # TODO: Implement auto-renewal toggle
        # 1. Get user's active subscription
        # 2. Update auto_renew setting
        # 3. Log preference change
        
        return {
            "success": True,
            "message": f"自动续费已{'开启' if enabled else '关闭'}"
        }
        
    except Exception as e:
        logger.error(f"Toggle auto-renew error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="更新自动续费设置失败"
        )