# backend/app/api/v1/translation.py
"""
Translation API endpoints

This module provides API endpoints for media translation tasks,
including file upload, processing status, and result retrieval.
"""

from fastapi import APIRouter, HTTPException, Depends, UploadFile, File, Form, status
from fastapi.security import HTTPBearer
from pydantic import BaseModel
from typing import Optional, List
import logging

from app.core.logging import get_logger

logger = get_logger(__name__)
router = APIRouter()
security = HTTPBearer()


# Pydantic models
class TranslationUploadRequest(BaseModel):
    """Translation upload request model."""
    file_type: str
    original_filename: Optional[str] = None
    metadata: Optional[dict] = None


class TranslationStartRequest(BaseModel):
    """Translation start request model."""
    task_id: str
    language: Optional[str] = "zh-CN"
    provider: Optional[str] = "alibaba"


class TranslationResponse(BaseModel):
    """Translation response model."""
    id: int
    task_id: str
    file_type: str
    original_filename: Optional[str]
    file_size: Optional[int]
    duration_seconds: Optional[float]
    processing_status: str
    processing_error: Optional[str]
    transcribed_text: Optional[str]
    confidence_score: Optional[float]
    word_count: Optional[int]
    created_at: str
    updated_at: str


class TranslationListResponse(BaseModel):
    """Translation list response model."""
    success: bool
    data: List[TranslationResponse]
    total: int
    page: int
    limit: int


class TranslationStatusResponse(BaseModel):
    """Translation status response model."""
    task_id: str
    processing_status: str
    progress_percentage: int
    processing_started_at: Optional[str]
    processing_completed_at: Optional[str]
    estimated_completion: Optional[str]
    error_message: Optional[str]


@router.post("/upload", response_model=dict)
async def upload_file(
    file: UploadFile = File(...),
    file_type: str = Form(...),
    current_user: dict = Depends(security)
):
    """
    Upload media file for translation.
    
    This endpoint handles file upload, validation, and creates
    a translation task for processing.
    """
    try:
        # TODO: Implement file upload logic
        # 1. Validate file type and size
        # 2. Save file to storage
        # 3. Create translation task
        # 4. Return task ID and upload info
        
        # Mock response
        mock_task_id = "mock_task_id_12345"
        
        return {
            "success": True,
            "data": {
                "task_id": mock_task_id,
                "file_info": {
                    "filename": file.filename,
                    "file_size": 1024000,  # Mock size
                    "file_type": file_type,
                    "upload_url": f"/uploads/{mock_task_id}/{file.filename}"
                }
            },
            "message": "文件上传成功"
        }
        
    except Exception as e:
        logger.error(f"File upload error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="文件上传失败，请稍后重试"
        )


@router.post("/start", response_model=dict)
async def start_translation(
    request: TranslationStartRequest,
    current_user: dict = Depends(security)
):
    """
    Start translation processing for uploaded file.
    
    This endpoint initiates the speech recognition process
    for the previously uploaded file.
    """
    try:
        # TODO: Implement translation start logic
        # 1. Validate task exists and belongs to user
        # 2. Check user quota availability
        # 3. Queue translation task
        # 4. Update task status to processing
        # 5. Return processing info
        
        return {
            "success": True,
            "data": {
                "task_id": request.task_id,
                "status": "processing",
                "estimated_duration": 120,  # seconds
                "provider": request.provider,
                "language": request.language
            },
            "message": "翻译已开始处理"
        }
        
    except Exception as e:
        logger.error(f"Translation start error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="翻译开始失败，请稍后重试"
        )


@router.get("/{task_id}/status", response_model=TranslationStatusResponse)
async def get_translation_status(
    task_id: str,
    current_user: dict = Depends(security)
):
    """
    Get translation processing status.
    
    This endpoint returns the current status of a translation task
    including progress and estimated completion time.
    """
    try:
        # TODO: Implement status check logic
        # 1. Validate task exists and belongs to user
        # 2. Get current processing status
        # 3. Calculate progress percentage
        # 4. Estimate completion time
        # 5. Return status information
        
        # Mock response
        return TranslationStatusResponse(
            task_id=task_id,
            processing_status="processing",
            progress_percentage=45,
            processing_started_at="2024-01-01T12:00:00Z",
            processing_completed_at=None,
            estimated_completion="2024-01-01T12:02:00Z",
            error_message=None
        )
        
    except Exception as e:
        logger.error(f"Translation status error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取翻译状态失败"
        )


@router.get("/{task_id}/result", response_model=dict)
async def get_translation_result(
    task_id: str,
    current_user: dict = Depends(security)
):
    """
    Get completed translation result.
    
    This endpoint returns the final transcription result
    when the translation task is completed.
    """
    try:
        # TODO: Implement result retrieval logic
        # 1. Validate task exists and belongs to user
        # 2. Check if translation is completed
        # 3. Get transcription result
        # 4. Return formatted result
        
        # Mock response
        return {
            "success": True,
            "data": {
                "task_id": task_id,
                "status": "completed",
                "transcribed_text": "这是语音识别的模拟结果文本。",
                "confidence_score": 0.95,
                "word_count": 12,
                "duration_seconds": 15.5,
                "processing_time": 18.2,
                "provider": "alibaba",
                "created_at": "2024-01-01T12:00:00Z",
                "completed_at": "2024-01-01T12:00:18Z"
            },
            "message": "翻译完成"
        }
        
    except Exception as e:
        logger.error(f"Translation result error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取翻译结果失败"
        )


@router.get("/history", response_model=TranslationListResponse)
async def get_translation_history(
    page: int = 1,
    limit: int = 20,
    status: Optional[str] = None,
    current_user: dict = Depends(security)
):
    """
    Get user's translation history.
    
    This endpoint returns a paginated list of user's
    previous translation tasks.
    """
    try:
        # TODO: Implement history retrieval logic
        # 1. Query user's translation tasks
        # 2. Apply pagination and filters
        # 3. Format response
        # 4. Return paginated results
        
        # Mock response
        mock_translations = [
            {
                "id": 1,
                "task_id": "mock_task_1",
                "file_type": "audio",
                "original_filename": "test.mp3",
                "file_size": 1024000,
                "duration_seconds": 15.5,
                "processing_status": "completed",
                "processing_error": None,
                "transcribed_text": "测试文本内容",
                "confidence_score": 0.95,
                "word_count": 4,
                "created_at": "2024-01-01T12:00:00Z",
                "updated_at": "2024-01-01T12:00:18Z"
            }
        ]
        
        return TranslationListResponse(
            success=True,
            data=mock_translations,
            total=1,
            page=page,
            limit=limit
        )
        
    except Exception as e:
        logger.error(f"Translation history error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取翻译历史失败"
        )


@router.delete("/{task_id}", response_model=dict)
async def delete_translation(
    task_id: str,
    current_user: dict = Depends(security)
):
    """
    Delete translation task and associated files.
    
    This endpoint permanently removes a translation task
    and all associated data/files.
    """
    try:
        # TODO: Implement deletion logic
        # 1. Validate task exists and belongs to user
        # 2. Delete associated files
        # 3. Delete translation record
        # 4. Delete usage records
        
        return {
            "success": True,
            "message": "翻译任务已删除"
        }
        
    except Exception as e:
        logger.error(f"Translation deletion error: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="删除翻译任务失败"
        )