# specs/001-wechat-media-translator/contracts/api.yaml
openapi: 3.0.3
info:
  title: WeChat Media Translator API
  description: RESTful API for WeChat mini-program media-to-text translation service
  version: 1.0.0
  contact:
    name: API Support
    email: support@wechat-translator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.wechat-translator.com/v1
    description: Production server
  - url: https://staging-api.wechat-translator.com/v1
    description: Staging server
paths:
...
  /translations/{translation_id}:
    get:
      summary: Get translation details
      description: Retrieve specific translation details and results
      tags:
        - Translation
      security:
        - BearerAuth: []
      parameters:
        - name: translation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Translation details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TranslationDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete translation
      description: Delete translation record and associated files
      tags:
        - Translation
      security:
        - BearerAuth: []
      parameters:
        - name: translation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Translation deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Subscription Management
...
  /subscriptions/current:
    get:
      summary: Get current subscription
      description: Retrieve user's current subscription details
      tags:
        - Subscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current subscription retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Subscription'
  /subscriptions/cancel:
    post:
      summary: Cancel subscription
      description: Cancel current subscription (effective at end of period)
      tags:
        - Subscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Subscription cancelled successfully"

  # Payment Webhooks
...
components:
  schemas:
...
    Translation:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440001"
        original_filename:
          type: string
          example: "voice_message.mp3"
        file_type:
          type: string
          enum: [voice_recording, audio_file, video_file, wechat_video, url_media]
          example: "voice_recording"
        file_size:
          type: integer
          example: 2048576
        file_duration:
          type: integer
          example: 45
        language:
          type: string
          example: "zh-CN"
        transcription_text:
          type: string
          example: "你好，我想预约一个明天的会议。"
        confidence_score:
          type: number
          example: 0.95
        processing_status:
          type: string
          enum: [queued, processing, completed, failed]
          example: "completed"
        processing_provider:
          type: string
          example: "alibaba_cloud"
        processing_time:
          type: integer
          example: 2500
        created_at:
          type: string
          format: date-time
          example: "2025-10-18T14:30:00Z"
        completed_at:
          type: string
          format: date-time
          example: "2025-10-18T14:32:33Z"
        expires_at:
          type: string
          format: date-time
          example: "2025-11-17T14:30:00Z"
    TranslationDetails:
      allOf:
        - $ref: '#/components/schemas/Translation'
        - type: object
          properties:
            file_url:
              type: string
              example: "https://cloud-storage.wechat.com/files/voice_001.mp3"
            source_type:
              type: string
              enum: [direct_upload, wechat_message, url_import]
              example: "wechat_message"
            error_message:
              type: string
              example: null
            tags:
              type: array
              items:
                type: string
              example: ["工作", "会议"]
...
    Subscription:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440002"
        tier:
          type: string
          enum: [basic, premium, enterprise]
          example: "basic"
        status:
          type: string
          enum: [active, expired, cancelled, pending]
          example: "active"
        start_date:
          type: string
          format: date-time
          example: "2025-10-18T00:00:00Z"
        end_date:
          type: string
          format: date-time
          example: "2025-11-18T00:00:00Z"
        auto_renew:
          type: boolean
          example: true
        price:
          type: integer
          example: 990
        currency:
          type: string
          example: "CNY"
        billing_cycle:
          type: string
          enum: [monthly, yearly]
          example: "monthly"
        created_at:
          type: string
          format: date-time
          example: "2025-10-18T10:30:00Z"
...
  responses:
...
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "unauthorized"
              message:
                type: string
                example: "Authentication required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "not_found"
              message:
                type: string
                example: "Requested resource not found"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...